Moved userService in nulp_elite
